Removing TODO (#1061)
