+++ quesma/table_resolver/table_resolver.go
@@ -7,6 +7,7 @@ import (
 	"fmt"
 	"quesma/clickhouse"
 	"quesma/elasticsearch"
+	"quesma/frontend_connectors"
 	"quesma/logger"
 	"quesma/quesma/config"
 	"quesma/quesma/recovery"
@@ -16,7 +17,7 @@ import (
 )
 
 type tableResolver interface {
-	resolve(indexPattern string) *Decision
+	resolve(indexPattern string) *frontend_connectors.Decision
 }
 
 // parsedPattern stores the parsed index pattern
@@ -31,17 +32,17 @@ type parsedPattern struct {
 
 type patternSplitter struct {
 	name     string
-	resolver func(pattern string) (parsedPattern, *Decision)
+	resolver func(pattern string) (parsedPattern, *frontend_connectors.Decision)
 }
 
 type basicResolver struct {
 	name     string
-	resolver func(part string) *Decision
+	resolver func(part string) *frontend_connectors.Decision
 }
 
 type decisionMerger struct {
 	name   string
-	merger func(decisions []*Decision) *Decision
+	merger func(decisions []*frontend_connectors.Decision) *frontend_connectors.Decision
 }
 
 // Compound resolver works in the following way:
@@ -54,14 +55,14 @@ type compoundResolver struct {
 	decisionMerger  decisionMerger
 }
 
-func (ir *compoundResolver) resolve(indexName string) *Decision {
+func (ir *compoundResolver) resolve(indexName string) *frontend_connectors.Decision {
 	input, decision := ir.patternSplitter.resolver(indexName)
 	if decision != nil {
 		decision.ResolverName = ir.patternSplitter.name
 		return decision
 	}
 
-	var decisions []*Decision
+	var decisions []*frontend_connectors.Decision
 	for _, part := range input.parts {
 		for _, resolver := range ir.decisionLadder {
 			decision := resolver.resolver(part)
@@ -81,9 +82,9 @@ func (ir *compoundResolver) resolve(indexName string) *Decision {
 // now we have a single config for both, but with different fields
 func getTargets(indexConf config.IndexConfiguration, pipeline string) []string {
 	switch pipeline {
-	case IngestPipeline:
+	case frontend_connectors.IngestPipeline:
 		return indexConf.IngestTarget
-	case QueryPipeline:
+	case frontend_connectors.QueryPipeline:
 		return indexConf.QueryTarget
 	default:
 		return []string{}
@@ -100,7 +101,7 @@ type pipelineResolver struct {
 	pipelineName string
 
 	resolver        tableResolver
-	recentDecisions map[string]*Decision
+	recentDecisions map[string]*frontend_connectors.Decision
 }
 
 type tableRegistryImpl struct {
@@ -118,13 +119,13 @@ type tableRegistryImpl struct {
 	conf              config.QuesmaConfiguration
 }
 
-func (r *tableRegistryImpl) Resolve(pipeline string, indexPattern string) *Decision {
+func (r *tableRegistryImpl) Resolve(pipeline string, indexPattern string) *frontend_connectors.Decision {
 	r.m.Lock()
 	defer r.m.Unlock()
 
 	res, exists := r.pipelineResolvers[pipeline]
 	if !exists {
-		return &Decision{
+		return &frontend_connectors.Decision{
 			IndexPattern: indexPattern,
 			Err:          fmt.Errorf("pipeline '%s' not found", pipeline),
 			Reason:       "Pipeline not found. This is a bug.",
@@ -194,7 +195,7 @@ func (r *tableRegistryImpl) updateIndexes() {
 	r.elasticIndexes = elasticIndexes
 	r.clickhouseIndexes = clickhouseIndexes
 	for _, res := range r.pipelineResolvers {
-		res.recentDecisions = make(map[string]*Decision)
+		res.recentDecisions = make(map[string]*frontend_connectors.Decision)
 	}
 }
 
@@ -223,7 +224,7 @@ func (r *tableRegistryImpl) Start() {
 	}()
 }
 
-func (r *tableRegistryImpl) RecentDecisions() []PatternDecisions {
+func (r *tableRegistryImpl) RecentDecisions() []frontend_connectors.PatternDecisions {
 	r.m.Lock()
 	defer r.m.Unlock()
 
@@ -242,12 +243,12 @@ func (r *tableRegistryImpl) RecentDecisions() []PatternDecisions {
 
 	sort.Strings(patterns)
 
-	var res []PatternDecisions
+	var res []frontend_connectors.PatternDecisions
 	for _, p := range patterns {
 
-		pd := PatternDecisions{
+		pd := frontend_connectors.PatternDecisions{
 			Pattern:   p,
-			Decisions: make(map[string]*Decision),
+			Decisions: make(map[string]*frontend_connectors.Decision),
 		}
 		for _, resolver := range r.pipelineResolvers {
 			if decision, ok := resolver.recentDecisions[p]; ok {
@@ -295,7 +296,7 @@ func NewTableResolver(quesmaConf config.QuesmaConfiguration, discovery clickhous
 	// TODO We should use the pipeline name as a key in the map.
 
 	ingestResolver := &pipelineResolver{
-		pipelineName: IngestPipeline,
+		pipelineName: frontend_connectors.IngestPipeline,
 
 		resolver: &compoundResolver{
 			patternSplitter: patternSplitter{
@@ -304,25 +305,25 @@ func NewTableResolver(quesmaConf config.QuesmaConfiguration, discovery clickhous
 			},
 			decisionLadder: []basicResolver{
 				{"kibanaInternal", resolveInternalElasticName},
-				{"disabled", makeIsDisabledInConfig(indexConf, IngestPipeline)},
+				{"disabled", makeIsDisabledInConfig(indexConf, frontend_connectors.IngestPipeline)},
 
-				{"singleIndex", res.singleIndex(indexConf, IngestPipeline)},
-				{"commonTable", res.makeCommonTableResolver(indexConf, IngestPipeline)},
+				{"singleIndex", res.singleIndex(indexConf, frontend_connectors.IngestPipeline)},
+				{"commonTable", res.makeCommonTableResolver(indexConf, frontend_connectors.IngestPipeline)},
 
-				{"defaultWildcard", makeDefaultWildcard(quesmaConf, IngestPipeline)},
+				{"defaultWildcard", makeDefaultWildcard(quesmaConf, frontend_connectors.IngestPipeline)},
 			},
 			decisionMerger: decisionMerger{
 				name:   "basicDecisionMerger",
 				merger: basicDecisionMerger,
 			},
 		},
-		recentDecisions: make(map[string]*Decision),
+		recentDecisions: make(map[string]*frontend_connectors.Decision),
 	}
 
-	res.pipelineResolvers[IngestPipeline] = ingestResolver
+	res.pipelineResolvers[frontend_connectors.IngestPipeline] = ingestResolver
 
 	queryResolver := &pipelineResolver{
-		pipelineName: QueryPipeline,
+		pipelineName: frontend_connectors.QueryPipeline,
 
 		resolver: &compoundResolver{
 			patternSplitter: patternSplitter{
@@ -332,23 +333,23 @@ func NewTableResolver(quesmaConf config.QuesmaConfiguration, discovery clickhous
 			decisionLadder: []basicResolver{
 				// checking if we can handle the parsedPattern
 				{"kibanaInternal", resolveInternalElasticName},
-				{"disabled", makeIsDisabledInConfig(indexConf, QueryPipeline)},
+				{"disabled", makeIsDisabledInConfig(indexConf, frontend_connectors.QueryPipeline)},
 
-				{"singleIndex", res.singleIndex(indexConf, QueryPipeline)},
-				{"commonTable", res.makeCommonTableResolver(indexConf, QueryPipeline)},
+				{"singleIndex", res.singleIndex(indexConf, frontend_connectors.QueryPipeline)},
+				{"commonTable", res.makeCommonTableResolver(indexConf, frontend_connectors.QueryPipeline)},
 
 				// default action
-				{"defaultWildcard", makeDefaultWildcard(quesmaConf, QueryPipeline)},
+				{"defaultWildcard", makeDefaultWildcard(quesmaConf, frontend_connectors.QueryPipeline)},
 			},
 			decisionMerger: decisionMerger{
 				name:   "basicDecisionMerger",
 				merger: basicDecisionMerger,
 			},
 		},
-		recentDecisions: make(map[string]*Decision),
+		recentDecisions: make(map[string]*frontend_connectors.Decision),
 	}
 
-	res.pipelineResolvers[QueryPipeline] = queryResolver
+	res.pipelineResolvers[frontend_connectors.QueryPipeline] = queryResolver
 	// update the state ASAP
 	res.updateState()
 	return res
